requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
huggingface-hub
